[Refactor] (test) Move the util object '_ZKNodePathUtils' to common module '_instance_value'.
